fix : Dto변환 Service계층에서 하도록 수정

+++ src/main/java/org/src/todo/dto/todo/TodoRequestDto.java
@@ -10,5 +10,5 @@ import org.src.todo.entity.Todo;
 @RequiredArgsConstructor
 public class TodoRequestDto {
     private final String contents;
-    private final int userId;
+    private final Long userId;
 }
